refactor: move stuff to mixin
